added admin verification

+++ src/app.ts
@@ -1,4 +1,5 @@
 import * as http from "http";
+import { AdminController } from "./controllers/AdminController";
 import { StudentController } from "./controllers/StudentsController";
 import { TeacherController } from "./controllers/TeachersController";
 import { getReqData } from "./utils";
@@ -30,6 +31,17 @@ const server = http.createServer(async (req, res) => {
     req.url.match(/\/api\/teachers\/([0-9a-zA-Z]+)/) &&
     req.method === "DELETE"
   ) {
+    if (!req.headers.authorization) {
+      res.writeHead(403);
+      res.end();
+      return;
+    }
+    if (!new AdminController().isAdmin(req.headers.authorization)) {
+      res.writeHead(401, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({ message: "the user does not have permission" }));
+      return;
+    }
+
     try {
       const id = req.url.split("/")[3];
       const message: string = await teachersController.deleteTeacher(id);
@@ -57,6 +69,17 @@ const server = http.createServer(async (req, res) => {
       res.end(JSON.stringify({ message: error }));
     }
   } else if (req.url === "/api/teachers" && req.method === "POST") {
+    if (!req.headers.authorization) {
+      res.writeHead(403);
+      res.end();
+      return;
+    }
+    if (!new AdminController().isAdmin(req.headers.authorization)) {
+      res.writeHead(401, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({ message: "the user does not have permission" }));
+      return;
+    }
+
     const newTeacherObj: string = (await getReqData(req)) as string;
     try {
       const teacher = await teachersController.createTeacher(
@@ -89,6 +112,17 @@ const server = http.createServer(async (req, res) => {
     req.url.match(/\/api\/students\/([0-9a-zA-Z]+)/) &&
     req.method === "DELETE"
   ) {
+    if (!req.headers.authorization) {
+      res.writeHead(403);
+      res.end();
+      return;
+    }
+    if (!new AdminController().isAdmin(req.headers.authorization)) {
+      res.writeHead(401, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({ message: "the user does not have permission" }));
+      return;
+    }
+
     try {
       const id = req.url.split("/")[3];
       const message: string = await studentsController.deleteStudent(id);
@@ -116,6 +150,17 @@ const server = http.createServer(async (req, res) => {
       res.end(JSON.stringify({ message: error }));
     }
   } else if (req.url === "/api/students" && req.method === "POST") {
+    if (!req.headers.authorization) {
+      res.writeHead(403);
+      res.end();
+      return;
+    }
+    if (!new AdminController().isAdmin(req.headers.authorization)) {
+      res.writeHead(401, { "Content-Type": "application/json" });
+      res.end(JSON.stringify({ message: "the user does not have permission" }));
+      return;
+    }
+
     const newStudentObj: string = await getReqData(req);
     try {
       const student = await studentsController.createStudent(
